+++ R/package.R
@@ -18,13 +18,15 @@
 #' 
 #' \bold{Flow Cytometry Data}
 #' 
+#' \code{\link{fcsprocessor}} Converts a single or a collection of objects of class 'fcs' to an object of class 'data.frame' readable by \code{\link{gating}}, \code{\link{rrs}}, and \code{\link{lotrrs}} functions. This function is used to compute the \code{\link{randCyto}} data.
+#' 
 #' \code{\link{randCyto}} A sample dataset containing information about flow cytometry data with two binary categorical variables. The data are a random subset of the 'extdata' data in the 'flowWorkspaceData' package found on Bioconductor \url{http://bioconductor.org/packages/release/data/experiment/html/flowWorkspaceData.html} and formatted for gateR input.
 #' 
 #' @name gateR-package
 #' @aliases gateR-package gateR
 #' @docType package
 #' 
-#' @section Dependencies: The 'gateR' package relies heavily upon \code{\link{sparr}}, \code{\link{spatstat.geom}}, and \code{\link{raster}}. For a two-level comparison, the spatial relative risk function uses the \code{\link[sparr]{risk}} function. The calculation of a Bonferroni correction for multiple testing accounting for the spatial correlation of the estimated surface uses the \code{\link[SpatialPack]{modified.ttest}} function. Basic visualizations rely on the \code{\link[fields]{image.plot}} function.
+#' @section Dependencies: The 'gateR' package relies heavily upon \code{\link{sparr}}, \code{\link{spatstat.geom}}, and \code{\link{raster}}. For a two-level comparison, the spatial relative risk function uses the \code{\link[sparr]{risk}} function. The calculation of a Bonferroni correction for multiple testing accounting for the spatial correlation of the estimated surface uses the \code{\link[SpatialPack]{modified.ttest}} function. The \code{\link{fcsprocessor}} function relies on the \code{\link{flowCore}} package found on Bioconductor \url{http://bioconductor.org/packages/release/data/experiment/html/flowWorkspaceData.html}. Basic visualizations rely on the \code{\link[fields]{image.plot}} function.
 #' 
 #' @author Ian D. Buller\cr \emph{Occupational and Environmental Epidemiology Branch, Division of Cancer Epidemiology and Genetics, National Cancer Institute, National Institutes of Health, Rockville, Maryland, USA.}
 #' 
@@ -34,6 +36,7 @@
 NULL
 
 #' @importFrom fields image.plot
+#' @importFrom flowCore exprs getChannelMarker identifier read.flowSet
 #' @importFrom graphics close.screen par screen split.screen 
 #' @importFrom grDevices chull colorRampPalette dev.off png
 #' @importFrom lifecycle badge deprecate_warn deprecated is_present
@@ -43,6 +46,8 @@ NULL
 #' @importFrom SpatialPack modified.ttest
 #' @importFrom stats na.omit pnorm relevel
 #' @importFrom tibble add_column
+#' @importFrom tools file_path_sans_ext
+#' @importFrom utils read.csv
 #' @import spatstat.geom
 #' @import spatstat.core
 #' @import spatstat.linnet